+++ vasttools/query.py
@@ -53,7 +53,7 @@ from vasttools.survey import (
 from vasttools.source import Source
 from vasttools.utils import (
     filter_selavy_components, simbad_search, match_planet_to_field,
-    epoch12_user_warning, read_selavy
+    read_selavy
 )
 from vasttools.moc import VASTMOCS
 from forced_phot import ForcedPhot
@@ -398,9 +398,6 @@ class Query:
             self.query_df = None
 
         self.fields_found = False
-
-        # TODO: Remove warning in future release.
-        epoch12_user_warning()
 
     def _validate_settings(self) -> bool:
         """